Finish the move to lib/
