+++ lib/array_like.js
@@ -4,7 +4,7 @@
 
 'use strict';
 
-// CreateModelArray is a builder for the ModelArray class. 
+// CreateArrayLike is a builder for the ArrayLike class. 
 //
 // All builders take one parameter which is a map of services. This is
 // the dependency injection mechanism.  Builder must return a class.
@@ -12,12 +12,12 @@
 // of the builder (so it is not possible to extend builder classes)
 // but code within the class (such as the class constructor) can
 // safely acess the services.
-export function CreateModelArray(services) {
+export function CreateArrayLike(services) {
     const empty = [];
 
-    // ModelArray is an effective singleton -- specific classes are
+    // ArrayLike is an effective singleton -- specific classes are
     // implemented by using the `ofConstructor` sub method.
-    return class ModelArray {
+    return class ArrayLike {
         static forEach(data, cb) {
             const json = data || empty;
             if (!Array.isArray(json)) {
