+++ lib/sparse_array.js
@@ -68,7 +68,7 @@ export function CreateSparseArray(services) {
         splice(offset, before, after) {
             if (offset < 0 || count < 0) throw new Error("splice: invalid parameter");
 
-            const count = services.ModelArray.count(before);
+            const count = services.ArrayLike.count(before);
             const end = offset + count;
             const e = [];
             const append = elt => {
@@ -79,12 +79,12 @@ export function CreateSparseArray(services) {
             let seen = 0;
             this.forEach((elt, ii) => {
                 seen ++;
-                if (ii === offset) services.ModelArray.forEach(after, append);
+                if (ii === offset) services.ArrayLike.forEach(after, append);
                 if (ii < offset || ii >= end) append(elt);
             });
 
             if (seen < end) throw new Error("splice: invalid params");
-            if (seen === end) services.ModelArray.forEach(after, append);
+            if (seen === end) services.ArrayLike.forEach(after, append);
             return new SparseArray(make(e));
         }
