+++ test/array_like_test.js
@@ -4,22 +4,22 @@
 
 'use strict';
 
-import {CreateSparseArray} from '../client/model/sparse_array.js';
-import {CreateModelArray} from '../client/model/array.js';
+import {CreateSparseArray} from '../lib/sparse_array.js';
+import {CreateArrayLike} from '../lib/array_like.js';
 
 describe('Array tests', () => {
     const services = {};
     services.SparseArray = CreateSparseArray(services);
-    services.ModelArray = CreateModelArray(services);
+    services.ArrayLike = CreateArrayLike(services);
     
     it ('counts arrays and array-like objects', () => {
-        const zero1 = services.ModelArray.count();
-        const zero2 = services.ModelArray.count(null);
-        const zero3 = services.ModelArray.count("");
-        const zero4 = services.ModelArray.count([]);
-        const zero5 = services.ModelArray.count({"dot:encoding": "SparseArray"});
-        const zero6 = services.ModelArray.count({"dot:encoding": "SparseArray", "dot:encoded": []});
-        const zero7 = services.ModelArray.count(new services.SparseArray());
+        const zero1 = services.ArrayLike.count();
+        const zero2 = services.ArrayLike.count(null);
+        const zero3 = services.ArrayLike.count("");
+        const zero4 = services.ArrayLike.count([]);
+        const zero5 = services.ArrayLike.count({"dot:encoding": "SparseArray"});
+        const zero6 = services.ArrayLike.count({"dot:encoding": "SparseArray", "dot:encoded": []});
+        const zero7 = services.ArrayLike.count(new services.SparseArray());
 
         const expected = [0, 0, 0, 0, 0, 0, 0];
         const actual = [zero1, zero2, zero3, zero4, zero5, zero6, zero7];
@@ -27,13 +27,13 @@ describe('Array tests', () => {
             throw new Error("Mismatched: " + JSON.stringify(actual));
         }
 
-        const three1 = services.ModelArray.count([1, 2, "hello"]);
+        const three1 = services.ArrayLike.count([1, 2, "hello"]);
         const encoded = {
             "dot:encoding": "SparseArray",
             "dot:encoded": [3, "hello"],
         };
-        const three2 = services.ModelArray.count(encoded);
-        const three3 = services.ModelArray.count(new services.SparseArray(encoded));
+        const three2 = services.ArrayLike.count(encoded);
+        const three3 = services.ArrayLike.count(new services.SparseArray(encoded));
 
         if (three1 !== 3 || three2 !== 3 || three3 !== 3) {
             throw new Error("Unexpexted counts: " + [three1, three2, three3]);
@@ -70,7 +70,7 @@ describe('Array tests', () => {
         
         function toArray(json) {
             const r = [];
-            services.ModelArray.forEach(json, (elt, index) => (r[index] = elt));
+            services.ArrayLike.forEach(json, (elt, index) => (r[index] = elt));
             return r;
         }
     });
